+++ web/resources/lib/account-panel.ts
@@ -1,4 +1,4 @@
 import { Modal } from "@wikijump/components"
-import AccountPanel from "./components/AccountPanel.svelte"
+import AccountPanel from "./components/account/AccountPanel.svelte"
 
 export const AccountModal = new Modal(AccountPanel)
