+++ web/resources/lib/components/account/NotificationBell.svelte
