+++ web/resources/lib/components/auth/ConfirmPasswordForm.svelte
