+++ web/resources/lib/components/auth/ForgotPasswordForm.svelte
