+++ web/resources/lib/components/auth/FormError.svelte
