+++ web/resources/lib/components/auth/RegisterForm.svelte
