+++ web/resources/lib/components/auth/ResendVerificationEmail.svelte
